feat: create component for looking up a provider
This removes the fact that provider access was not possible in the
template without having to inject a provider into your component manually.

Now, using `get-provider`, you can look up a provider and yield a reference
to it, useful if you want to pass that object into another component that
you do not want to modify.